+++ app/upgrades/latest/upgrades.go
@@ -9,6 +9,7 @@ import (
 	"github.com/archway-network/archway/app/keepers"
 	"github.com/archway-network/archway/app/upgrades"
 	callbackTypes "github.com/archway-network/archway/x/callback/types"
+	"github.com/archway-network/archway/x/cwfees"
 )
 
 // This upgrade handler is used for all the current changes to the protocol
@@ -32,6 +33,7 @@ var Upgrade = upgrades.Upgrade{
 	StoreUpgrades: storetypes.StoreUpgrades{
 		Added: []string{
 			callbackTypes.ModuleName,
+			cwfees.ModuleName,
 		},
 	},
 }