+++ x/cwfees/types/keys.go
@@ -1,0 +1,9 @@
+package types
+
+import "cosmossdk.io/collections"
+
+const ModuleName = "cwfees"
+
+var (
+	GrantersPrefix = collections.NewPrefix(0)
+)